Breaking change: require exclusive access
The fw_cfg device is a shared resource: there’s only one per QEMU guest.
And it is stateful: when happens when accessing an I/O port depends
on previous accesses.
This means that concurrent access would behave incorectly
(though probably not cause memory unsafety, as long as DMA is not used).

To fix this, require of users (through `unsafe fn new`) to only have
one `FwCfg` value at a time, and change methods to take `&mut self`
instead of `&self`.

This is a breaking API change, update the version number accordingly.